Moved data file to new data folder which seems a more appropriate place, and imported facilitiesData into the form.
